@@ -1,8 +1,6 @@
 module github.com/apache/skywalking-eyes
 
-go 1.23.0
-
-toolchain go1.23.6
+go 1.25.3
 
 require (
 	github.com/Masterminds/sprig/v3 v3.2.3
